requirements.txt
streamlit
openai
python-dotenv
pypdf
pdfplumber
pytesseract
Pillow